+++ analytics/user_activity.py
@@ -6,7 +6,7 @@ import os
 from imgur import upload
 
 
-def user_activity(reddit_user, save_path=''):  # TODO: Implement timezone
+def user_activity(reddit_user, save_path='', debug=False):  # TODO: Implement timezone
     # Set to grab a certain number of things from reddit...reddit wont return more than 1000
     thing_limit = 100
     # Dictionary that contains the indexes to access information in the datetime.tuple()
@@ -60,9 +60,19 @@ def user_activity(reddit_user, save_path=''):  # TODO: Implement timezone
     # Saves a png of the generated report
     file_name = os.path.join(save_path + str(reddit_user.name) + '_user_activity.png')
     plt.savefig(file_name)
-    image_link = upload.upload_image(file_name)
 
     # Remove local copy of png
-    os.remove(file_name)
+    if not debug:
+        image_link = upload.upload_image(file_name)
+        os.remove(file_name)
+    else:
+        image_link = ''
 
     return image_link
+
+if __name__ == '__main__':
+    import praw
+    from bot.settings import user_agent
+    client = praw.Reddit(user_agent)
+    reddit_u = client.get_redditor('giantmatt')
+    user_activity(reddit_u, debug=True)
