+++ analytics/word_cloud.py
@@ -4,7 +4,7 @@ from imgur import upload
 import os
 
 
-def word_cloud(reddit_user, save_path=''):
+def word_cloud(reddit_user, save_path='', debug=False):
     # Set to grab a certain number of things from reddit...reddit wont return more than 1000
     thing_limit = 1000
 
@@ -29,9 +29,19 @@ def word_cloud(reddit_user, save_path=''):
     # Saves a png of the generated report
     file_name = os.path.join(save_path + reddit_user.name + '_word_cloud.png')
     plt.savefig(file_name)
-    image_link = upload.upload_image(file_name)
 
     # Remove local copy of png
-    os.remove(file_name)
+    if not debug:
+        image_link = upload.upload_image(file_name)
+        os.remove(file_name)
+    else:
+        image_link = ''
 
     return image_link
+
+if __name__ == '__main__':
+    import praw
+    from bot.settings import user_agent
+    client = praw.Reddit(user_agent)
+    reddit_u = client.get_redditor('giantmatt')
+    word_cloud(reddit_u, debug=True)
